make the RichText configurable
